fix(preview): feedback
