--- src/elements/common/async.js
@@ -1,3 +1,0 @@
-const importV4Annotations = () => import(/* webpackChunkName: "box-annotations" */ 'box-annotations');
-
-export default importV4Annotations;
