+++ src/elements/common/current-user/index.js.flow
@@ -1,0 +1,8 @@
+/**
+ * Flowtype definitions for index.ts
+ * Generated by Flowgen from a Typescript Definition
+ * Flowgen v1.10.0
+ */
+
+declare export { default } from "./withCurrentUser";
+declare export { default as withCurrentUser } from "./withCurrentUser";
