+++ src/elements/common/current-user/index.ts
@@ -1,0 +1,2 @@
+export { default } from './withCurrentUser';
+export { default as withCurrentUser } from './withCurrentUser';
