+++ src/elements/common/flowTypes.js
@@ -1,4 +1,5 @@
 // @flow
+import type { MessageDescriptor } from 'react-intl';
 import {
     VIEW_MODE_GRID,
     VIEW_MODE_LIST,
@@ -54,14 +55,31 @@ type ModalOptions = {
     overlayClassName: string,
 };
 
+type MaskError = {
+    errorHeader: MessageDescriptor,
+    errorSubHeader?: MessageDescriptor,
+};
+
+type InlineError = {
+    content: MessageDescriptor,
+    title: MessageDescriptor,
+};
+
+type Errors = {
+    error?: MessageDescriptor,
+    inlineError?: InlineError,
+    maskError?: MaskError,
+};
+
 export type {
-    ViewMode,
+    AdditionalVersionInfo,
+    Alignment,
+    ElementOrigin,
     ErrorType,
+    Errors,
     GetAvatarUrlCallback,
     GetProfileUrlCallback,
-    AdditionalVersionInfo,
-    ElementOrigin,
-    Alignment,
     ModalOptions,
     Page,
+    ViewMode,
 };
